refactor: 인터페이스와 구현체를 분리한 부분을 다시 통합
