+++ src/test/java/racingcar/repository/PlayerRepositoryTest.java
@@ -10,23 +10,23 @@ import org.springframework.boot.test.context.SpringBootTest;
 import racingcar.domain.CarGroup;
 
 @SpringBootTest
-class PlayerRepositoryImplTest {
+class PlayerRepositoryTest {
 
     @Autowired
-    private PlayerRepositoryImpl playerRepositoryImpl;
+    private PlayerRepository playerRepository;
 
     @Autowired
-    private RacingGameRepositoryImpl racingGameRepositoryImpl;
+    private RacingGameRepository racingGameRepository;
 
     @DisplayName("플레이어 저장")
     @Test
     void save() {
         // given
         CarGroup carGroup = new CarGroup("저문,헤나");
-        int racingGameId = racingGameRepositoryImpl.save("저문,헤나", 10);
+        int racingGameId = racingGameRepository.save("저문,헤나", 10);
 
         // when
-        boolean isSaved = playerRepositoryImpl.save(carGroup, racingGameId);
+        boolean isSaved = playerRepository.save(carGroup, racingGameId);
 
         // then
         assertThat(isSaved).isTrue();
